Remove test files and those tests that depend on them.
The test files exceed github's file size limit. The tests have been
removed until an alternative solution is found, or smaller test files are available.
